mood-tagger/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
transformers
Pillow
redis[hiredis]